backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "todo-backend"
version = "0.1.0"
description = "Backend for Todo application"
authors = []
requires-python = ">=3.11"
dependencies = [
    "Django>=5.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.1",
    "dj-database-url>=2.1.0",
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.9",
    "pytest>=8.3.4",
    "pytest-django>=4.7.0",
    "factory-boy>=3.3.0",
]

[tool.setuptools]
packages = ["backend", "todos"]

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = ["**/migrations/**"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
testpaths = ["todos/tests"]
